requirements.txt
pyodbc>=4.0
pandas>=2.0
sqlalchemy>=2.0
openpyxl>=3.1
python-dotenv>=1.0
tabulate>=0.9
rich>=13.0
plotly>=6.5
